# zstd: Don't allocate dataStorage when using byteBuf