@klauspost 2023-01-16
@greatroar Could you check https://github.com/klauspost/compress/pull/741 - I can't see how that would have regressions.